Adding a few fixes for minifloat around some less tested code paths

+++ _source/res/minifloat/minifloat/src/minifloat/FP8.h
@@ -16,7 +16,7 @@ float fp8_to_f32(uint8_t x);
 uint8_t f16_to_fp8(uint16_t x);
 inline uint16_t fp8_to_f16(uint8_t x) { return (uint16_t)x << 8; }
 
-inline uint8_t half_to_fp8(uint8_t x) { return f16_to_fp8(x); }
+inline uint8_t half_to_fp8(uint16_t x) { return f16_to_fp8(x); }
 inline uint16_t fp8_to_half(uint8_t x) { return fp8_to_f16(x); }
 
 // How to handle a value that exceeds the maximum magnitude.
@@ -40,9 +40,9 @@ float e4m3_to_f32(uint8_t x);
 // 1.5.2
 uint8_t f32_to_e5m2(float x, enum OFP8_SatMode sat_mode);
 inline float e5m2_to_f32(uint8_t x) { return fp8_to_f32(x); }
-uint8_t f16_to_e5m2(uint8_t x, enum OFP8_SatMode sat_mode);
+uint8_t f16_to_e5m2(uint16_t x, enum OFP8_SatMode sat_mode);
 inline uint16_t e5m2_to_f16(uint8_t x) { return fp8_to_f16(x); }
-inline uint8_t half_to_e5m2(uint8_t x, enum OFP8_SatMode sat_mode) {
+inline uint8_t half_to_e5m2(uint16_t x, enum OFP8_SatMode sat_mode) {
   return f16_to_e5m2(x, sat_mode);
 }
 inline uint16_t e5m2_to_half(uint8_t x) { return fp8_to_f16(x); }
